docs: npm run docusaurus docs:version v1alpha5
In preparation for v1alpha6 where we're designing a generic Command task
for use in generators, transformers, and validators, we need to create
permalinks for the existing v1alpha5 schemas.

This patch copies the current documentation to the v1alpha5 versioned
docs links and establishes v1alpha6 as the current version.

See: https://github.com/holos-run/holos/pull/413 for context.