render: discriminate components on typemeta.yaml
Previously, holos loaded the full CUE Instance to discriminate against
the APIVersion and Kind.  This is a problem because we need to know the
api version to know what tags to inject.  For example, v1alpha6 needs to
have the build context injected but v1alpha5 does not.

This patch fixes the problem by changing holos render component to look
for a `typemeta.yaml` file in the component directory.  The command line
tool discriminates on the kind and apiVersion specified in this file
prior to building the CUE instance.

Result:

We're able to inject version specific cue tags into the CUE instance.

Note previously the build context was passed to CUE by writing a file to
the filesystem and using the embed functionality.  We cannot proceed
with this embed approach because the filesystem writes are not safe for
concurrent use.  Therefore, we inject the build context as a tag which
is safe for concurrent use.